refactor(frontend): Extract types of Token UI to separate modules
