+++ src/frontend/src/lib/types/token-group.ts
@@ -3,8 +3,6 @@ import type {
 	TokenGroupPropSchema,
 	TokenGroupSchema
 } from '$lib/schema/token-group.schema';
-import type { TokenFinancialData, TokenUi } from '$lib/types/token';
-import type { NonEmptyArray } from '$lib/types/utils';
 import type * as z from 'zod';
 
 export type TokenGroupId = z.infer<typeof TokenGroupIdSchema>;
@@ -12,12 +10,3 @@ export type TokenGroupId = z.infer<typeof TokenGroupIdSchema>;
 export type TokenGroupData = z.infer<typeof TokenGroupSchema>;
 
 export type TokenGroup = z.infer<typeof TokenGroupPropSchema>;
-
-export type TokenUiGroup = {
-	id: TokenGroupId;
-	decimals: number;
-	groupData: TokenGroupData;
-	tokens: NonEmptyArray<TokenUi>;
-} & TokenFinancialData;
-
-export type TokenUiOrGroupUi = { token: TokenUi } | { group: TokenUiGroup };
