+++ src/frontend/src/lib/types/token.ts
@@ -55,10 +55,3 @@ export interface TokenFinancialData {
 	balance?: Exclude<OptionBalance, undefined>;
 	usdBalance?: number;
 }
-
-export type TokenUi<T extends Token = Token> = T & TokenFinancialData;
-
-export type TokenUiGroupable<T extends Token = Token> = Omit<TokenUi<T>, 'groupData'> &
-	Required<Pick<TokenUi<T>, 'groupData'>>;
-
-export type OptionTokenUi = Option<TokenUi>;
